pyproject.toml
[project]
name = "netmind"
version = "0.1.0"
description = "Dual-interface TCP Debugger and MCP Server with Hamlib support"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "mcp>=0.1.0",
    "fastmcp>=0.2.0",
    "websockets>=12.0",
    "jinja2>=3.1.0",
    "pydantic>=2.6.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-playwright>=0.4.0",
    "httpx>=0.27.0"
]

[project.scripts]
netmind-mcp = "netmind.mcp_stdio:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
filterwarnings = [
    "ignore:websockets.legacy is deprecated:DeprecationWarning",
    "ignore:websockets.server.WebSocketServerProtocol is deprecated:DeprecationWarning",
    "ignore:remove second argument of ws_handler:DeprecationWarning",
    "ignore:The `name` is not the first parameter anymore:DeprecationWarning"
]